+++ keanu-project/src/test/java/io/improbable/keanu/vertices/dbl/nonprobabilistic/operators/binary/SliceVertexTest.java
@@ -110,7 +110,7 @@ public class SliceVertexTest {
         DoubleVertex alpha = new UniformVertex(0, 10);
         alpha.setValue(DoubleTensor.create(new double[]{10, 15, 20, 25, 30, 35}, 3, 2));
 
-        MatrixMultiplicationVertex N = m.matrixMultiply(alpha);
+        DoubleVertex N = m.matrixMultiply(alpha);
 
         SliceVertex sliceN = new SliceVertex(N, dim, ind);
 
@@ -138,7 +138,7 @@ public class SliceVertexTest {
         DoubleVertex alpha = new UniformVertex(0, 10);
         alpha.setValue(DoubleTensor.create(new double[]{10, 15, 20, 25}, 2, 2));
 
-        MatrixMultiplicationVertex N = m.matrixMultiply(alpha);
+        DoubleVertex N = m.matrixMultiply(alpha);
 
         SliceVertex sliceN = new SliceVertex(N, 1, 1);